requirements.txt
openai
git+https://github.com/openai/swarm.git
portkey-ai
python-dotenv
